Revert "Press release (#47)"
This reverts commit 6cc3bf1d0e1da0a62911be1f1c55e6bbfc8c76d5.
