+++ src/main.ts
@@ -1,12 +1,10 @@
 // src/main.ts
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
-import router from './router'
 import App from './App.vue'
 
 import './assets/styles/main.scss'
 
 const app = createApp(App)
 app.use(createPinia())
-app.use(router)
 app.mount('#app')
