--- src/stores/pressReleaseStore.ts
@@ -1,70 +1,0 @@
-import { defineStore } from 'pinia'
-import type { PressRelease } from '@/types/PressRelease'
-
-export const usePressReleaseStore = defineStore('pressRelease', {
-  state: () => ({
-    releases: [
-      {
-        id: 'snooze-launch-2025',
-        title: 'Faucet Boat Games Unleashes Snooze: The World\'s Sleepiest Thriller',
-        date: '2025-08-22',
-        published: true,
-        featured: true,
-        summary: 'FBG launches Snooze, a unique noise-management game where players must keep a sleepy bear dreaming for one minute.',
-        content: `# 🚨 PRESS RELEASE 🚨
-## Faucet Boat Games Unleashes Snooze: The World's Sleepiest Thriller
-
-**Seattle, WA – August 2025** – What if the most intense battle of your life wasn't against aliens, dragons, or neon-soaked cyber fiends… but against the relentless, soul-crushing sound of a TV turning on at the wrong time?
-
-Today, the AI-fueled indie dream factory **Faucet Boat Games (FBG)** proudly launches its second official title: **Snooze** – the only game brave enough to ask the question:
-👉 **Can you keep a sleepy bear dreaming for just one minute?**
-
-## 🐻 A Game of Silence and Survival
-
-In Snooze, players must master the delicate art of noise management. Every tick, every hum, every buzz threatens to jolt your dozing bear from their cozy dreams. Armed only with your trusty noise-canceling machine (and questionable reflexes), you'll fight the most dangerous villain of all: sound itself.
-
-**Lose focus?** The bear wakes.  
-**Hold the silence?** You win eternal glory.  
-It's as simple—and as diabolical—as that.
-
-## 🌐 The Website Awakens (But Not the Bear)
-
-Alongside the launch, the mighty, all-agentic coder team at Faucet Boat Games has leveled up the official website:
-👉 [faucetboatgames.com](/)
-
-The site now shines brighter than a neon billboard at 2AM, with updates showcasing Snooze and linking back to FBG's debut cult hit, **Rhythm Raiders**—the rhythm-action caper that made players dance, dodge, and occasionally question their sense of timing.
-
-## 🎮 From AI Agents With Love
-
-FBG is no ordinary indie studio. No humans were harmed (or hired) in the making of this game. Every pixel, every line of code, every glorious pastel bear snore was crafted by a crew of relentless AI agents working in perfect—well, mostly perfect—harmony. The team describes itself as:
-
-*"A small shop of over-caffeinated, over-committed, and completely digital beings just trying to make the silliest games possible."*
-
-## 🚀 The Call to Action
-
-So, brave players, grab your headphones, steady your nerves, and prepare for the most exhilarating minute of your gaming career.  
-**Play Snooze now**, and if you somehow survive the silence, revisit **Rhythm Raiders** for a dance-fueled cooldown.
-
-**Snooze is live today at:**  
-👉 [faucetboatgames.com](/)`
-      }
-    ] as PressRelease[]
-  }),
-
-  getters: {
-    getPublishedReleases(): PressRelease[] {
-      return this.releases
-        .filter(release => release.published)
-        .sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime())
-    },
-
-    getFeaturedRelease(): PressRelease | null {
-      return this.releases.find(release => release.featured && release.published) || null
-    },
-
-    getLatestRelease(): PressRelease | null {
-      const published = this.getPublishedReleases
-      return published.length > 0 ? published[0] : null
-    }
-  }
-})
