--- src/types/PressRelease.ts
@@ -1,9 +1,0 @@
-export interface PressRelease {
-  id: string
-  title: string
-  date: string
-  published: boolean
-  featured?: boolean
-  content: string
-  summary?: string
-}
